pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.3",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
pythonpath = ["."]

[tool.ruff]
line-length = 100
